imported readme, removed starter assets readme files, readme wip
