+++ Basic/ClientDriven/Assets/Readme.asset
@@ -1,0 +1,42 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5ce0e387261ddc34ba81ca7f2f1f7f73, type: 3}
+  m_Name: Readme
+  m_EditorClassIdentifier: 
+  icon: {fileID: 2800000, guid: 9b46b2c2f21deb843b7ac9242d5f13e6, type: 3}
+  title: 'ClientDriven: A Bitesize Sample'
+  sections:
+  - heading: ClientDriven
+    text: 'ClientDriven is a bitesize sample designed to demonstrate responsive character
+      movements even under unfavourable network conditions. The sample leverages
+      Netcode for GameObject''s (Netcode) ClientNetworkTransform component and moves
+      the owning player client-side, client authoritatively. To learn more about
+      the sample, see the '
+    linkText: ClientDriven bitesize sample documentation.
+    url: https://docs-multiplayer.unity3d.com/netcode/current/learn/bitesize/bitesize-clientdriven/index.html
+  - heading: 
+    text: 
+    linkText: 
+    url: 
+  - heading: 
+    text: 
+    linkText: 
+    url: 
+  - heading: 
+    text: 
+    linkText: 
+    url: 
+  - heading: 
+    text: 
+    linkText: 
+    url: 
+  loadedLayout: 1
