+++ Basic/ClientDriven/Assets/Readme.asset.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 4181bfe715a83e843b7538d750dc57cc
-timeCreated: 1484217493
-licenseType: Store
+guid: 13de520ba66bd2e41b1019e0b6650e7f
 NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
