+++ Basic/ClientDriven/Assets/Scripts/Readme.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 65cc5574a07e76b4fb8e75a3f82b12fe
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
