+++ Basic/ClientDriven/Assets/Scripts/Readme/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 854341667afcba7499b57c898b0c5438
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
