+++ Basic/ClientDriven/Assets/Scripts/Readme/Editor/ReadmeEditor.cs.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 476cc7d7cd9874016adc216baab94a0a
-timeCreated: 1484146680
-licenseType: Store
+guid: e071f34aab1e7c245b3c500b5e8bd8eb
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
