+++ Basic/ClientDriven/Assets/Scripts/Readme/Readme.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5ce0e387261ddc34ba81ca7f2f1f7f73
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
