--- Basic/ClientDriven/Assets/StarterAssets/TutorialInfo/Scripts/Readme.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using UnityEngine;
-
-public class Readme : ScriptableObject {
-	public Texture2D icon;
-	public string title;
-	public Section[] sections;
-	public bool loadedLayout;
-	
-	[Serializable]
-	public class Section {
-		public string heading, text, linkText, url;
-	}
-}
