--- Basic/ClientDriven/Assets/StarterAssets/TutorialInfo/Scripts/Readme.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: fcf7219bab7fe46a1ad266029b2fee19
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences:
-  - icon: {fileID: 2800000, guid: d4743ba2e2a59f946b2125c074582ce7, type: 3}
-  executionOrder: 0
-  icon: {fileID: 2800000, guid: a186f8a87ca4f4d3aa864638ad5dfb65, type: 3}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
